# web/frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOMO Study Agent Web UI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/study/future": {
      "get": {
        "tags": [
          "study"
        ],
        "summary": "Get Future",
        "description": "获取未来 N 天的学习计划。",
        "operationId": "get_future_api_study_future_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_FutureItemsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/study/process-future": {
      "post": {
        "tags": [
          "study"
        ],
        "summary": "Process Future",
        "description": "触发未来计划处理，立即返回 task_id。",
        "operationId": "process_future_api_study_process_future_post",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TaskRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "UNKNOWN"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiError"
      },
      "ApiResponse_FutureItemsResponse_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FutureItemsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[FutureItemsResponse]"
      },
...
      "ApiResponse_TaskRunResponse_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[TaskRunResponse]"
      },
...
      "FutureItemsResponse": {
        "properties": {
          "days": {
            "type": "integer",
            "title": "Days",
            "default": 7
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TodayItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "FutureItemsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskRunResponse": {
        "properties": {
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "TaskRunResponse"
      },
...
      "TodayItem": {
        "properties": {
          "voc_id": {
            "type": "string",
            "title": "Voc Id"
          },
          "voc_spelling": {
            "type": "string",
            "title": "Voc Spelling"
          },
          "voc_meanings": {
            "type": "string",
            "title": "Voc Meanings",
            "default": ""
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count",
            "default": 0
          },
          "familiarity_short": {
            "type": "number",
            "title": "Familiarity Short",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "voc_id",
          "voc_spelling"
        ],
        "title": "TodayItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}